@@ -9,17 +9,18 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.4.2</version>
+        <version>3.3.3</version>
     </parent>
 
     <name>Grading Server</name>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <jackson-bom.version>2.16.0</jackson-bom.version>
         <junit-jupiter.version>5.9.3</junit-jupiter.version>
-        <hibernate.version>5.6.15.Final</hibernate.version>
+        <hibernate.version>6.2.31.Final</hibernate.version>
+        <h2.version>1.4.200</h2.version>
     </properties>
 
     <dependencies>
@@ -43,28 +44,28 @@
         <dependency>
             <groupId>io.pebbletemplates</groupId>
             <artifactId>pebble-spring-boot-starter</artifactId>
-            <version>3.1.6</version>
+            <version>3.2.2</version>
         </dependency>
 
         <dependency>
             <groupId>org.gitlab4j</groupId>
             <artifactId>gitlab4j-api</artifactId>
-            <version>5.5.0</version>
+            <version>6.0.0-rc.5</version>
         </dependency>
         <dependency>
             <groupId>org.eclipse.jgit</groupId>
             <artifactId>org.eclipse.jgit</artifactId>
-            <version>6.9.0.202403050737-r</version>
+            <version>7.0.0.202409031743-r</version>
         </dependency>
         <dependency>
             <groupId>org.ocpsoft.prettytime</groupId>
             <artifactId>prettytime</artifactId>
-            <version>5.0.8.Final</version>
+            <version>5.0.9.Final</version>
         </dependency>
         <dependency>
             <groupId>io.hypersistence</groupId>
-            <artifactId>hypersistence-utils-hibernate-52</artifactId>
-            <version>3.7.3</version>
+            <artifactId>hypersistence-utils-hibernate-62</artifactId>
+            <version>3.8.2</version>
         </dependency>
         <dependency>
             <groupId>com.h2database</groupId>
@@ -78,12 +79,12 @@
         <dependency>
             <groupId>ch.trick17.jtt</groupId>
             <artifactId>grader</artifactId>
-            <version>3.0.0-SNAPSHOT</version>
+            <version>3.1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>ch.trick17.jtt</groupId>
             <artifactId>test-suite-grader</artifactId>
-            <version>3.0.0-SNAPSHOT</version>
+            <version>3.1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
