@@ -1,7 +1,9 @@
 package ch.trick17.gradingserver.model;
 
-import javax.persistence.Embeddable;
-import javax.persistence.ManyToMany;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.ManyToMany;
+
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
@@ -14,6 +16,7 @@ public class ProjectConfig {
     private static final String PACKAGE_PATTERN = "[a-zA-Z_][a-zA-Z0-9_]*(\\.[a-zA-Z_][a-zA-Z0-9_]*)*";
 
     private String projectRoot;
+    @Column(columnDefinition = "integer")
     private ProjectStructure structure;
     private String packageFilter;       // null means no filter
     @ManyToMany
